Move script_run_context out of scriptrunner module
