+++ lib/streamlit/runtime/fragment.py
@@ -30,8 +30,8 @@ from streamlit.error_util import handle_uncaught_app_exception
 from streamlit.errors import FragmentHandledException, FragmentStorageKeyError
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
 from streamlit.runtime.metrics_util import gather_metrics
+from streamlit.runtime.script_run_context import get_script_run_ctx
 from streamlit.runtime.scriptrunner.exceptions import RerunException, StopException
-from streamlit.runtime.scriptrunner.script_run_context import get_script_run_ctx
 from streamlit.time_util import time_to_seconds
 
 if TYPE_CHECKING:
