+++ lib/streamlit/runtime/state/query_params.py
@@ -21,6 +21,7 @@ from urllib import parse
 from streamlit.constants import EMBED_QUERY_PARAMS_KEYS
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
+from streamlit.runtime.script_run_context import get_script_run_ctx
 
 if TYPE_CHECKING:
     from _typeshed import SupportsKeysAndGetItem
@@ -134,9 +135,6 @@ class QueryParams(MutableMapping[str, str]):
         return str(self._query_params)
 
     def _send_query_param_msg(self) -> None:
-        # Avoid circular imports
-        from streamlit.runtime.scriptrunner import get_script_run_ctx
-
         ctx = get_script_run_ctx()
         if ctx is None:
             return
@@ -189,9 +187,6 @@ class QueryParams(MutableMapping[str, str]):
         }
 
     def _ensure_single_query_api_used(self):
-        # Avoid circular imports
-        from streamlit.runtime.scriptrunner import get_script_run_ctx
-
         ctx = get_script_run_ctx()
         if ctx is None:
             return
