+++ lib/streamlit/user_info.py
@@ -17,7 +17,9 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Iterator, Mapping, NoReturn, Union
 
 from streamlit.errors import StreamlitAPIException
-from streamlit.runtime.scriptrunner import get_script_run_ctx as _get_script_run_ctx
+from streamlit.runtime.script_run_context import (
+    get_script_run_ctx as _get_script_run_ctx,
+)
 
 if TYPE_CHECKING:
     from streamlit.runtime.script_run_context import UserInfo
